% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalDT.R
\name{conditionalDT}
\alias{conditionalDT}
\title{conditional distribution of D|T}
\usage{
conditionalDT(data.predict.all, long_fit_all, survival_fit_all, l_i)
}
\arguments{
\item{data.predict.all}{This involves a collection of \code{data.frame} objects for
dynamic prediction, each corresponding to a distinct longitudinal outcome.
These data frames should contain the variables specified in \code{LongSubFixed}
and \code{LongSubRandom}. Utilizing a list structure
allows for the incorporation of multiple longitudinal outcomes,
each potentially following different measurement protocols.
In instances where all longitudinal outcomes are recorded at identical
time points across patients, a singular \code{data.frame} object may
be used in a \code{list}. It is presumed that each data frame is
structured in a long format.}

\item{long_fit_all}{Outputs from the model fitting process using the \code{nlme} package,
encompassing the results and parameters obtained from the analysis.}

\item{survival_fit_all}{Results and parameters generated from the model fitting
procedure, utilizing the \code{coxph} function. These outputs include the comprehensive
findings and variables derived from the analysis.}

\item{l_i}{A vector of time points to calculate the conditional probability.}
}
\value{
Probability matrices of competing risk event type D conditional on
survival outcome T.
}
\description{
This function computes the conditional probability density function of
competing risk event type D, given the survival time T.
}
\keyword{internal}
