% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinalSub.R
\name{longitudinalSub}
\alias{longitudinalSub}
\title{The process involves estimating parameters for a multivariate linear mixed-effects
model, which simultaneously analyzes multiple dependent variables that may be
correlated. This approach incorporates both fixed effects, which are consistent
across the population, and random effects, accounting for variations within
groups or subjects. By fitting this model, one can assess the influence of
predictor variables on several longitudinal outcomes while considering the inherent
variability in the data due to random effects.}
\usage{
longitudinalSub(data.fit.all, LongSubFixed, LongSubRandom)
}
\arguments{
\item{data.fit.all}{This process requires a set of \code{data.frame} objects
designated for model fitting, with each \code{data.frame} representing a
separate longitudinal outcome. These \code{data.frame} objects must include the
variables identified in \code{LongSubFixed} and \code{LongSubRandom}.
The use of a \code{list} arrangement facilitates the inclusion of
various longitudinal outcomes, which may adhere to different measurement protocols.
When all longitudinal outcomes are measured at the same time points for every patient,
a single \code{data.frame} object can be in a list.
It is assumed that every \code{data.frame} is organized in a long format.}

\item{LongSubFixed}{This refers to a collection of formulas detailing the
fixed effects portion for each longitudinal outcome. On the left side of each formula,
the response variable is defined, while the right side outlines
the fixed effect terms. Should only a single formula be provided—whether
as a list with one item or as a standalone formula—it is inferred that
a conventional univariate joint model is being constructed.}

\item{LongSubRandom}{A list of one-sided formulas that define the model for the
random effects of each longitudinal outcome.
The number of items in this \code{list} should match the length of \code{formLongFixed}.}
}
\value{
This structure comprises a list with four components.
The initial element, labeled \code{lfit}, consists of a collection of
outcomes from fitting multiple univariate linear mixed models,
where each entry within \code{lfit} corresponds to the results obtained
through the application of the \code{lme} function from the \code{nlme} package.
The second element is the estimated variance-covariance matrix derived from
the random effects in a multivariate linear mixed model.
The third and fourth elements, \code{LongSubFixed} and \code{LongSubRandom},
respectively, mirror the inputs provided to the model.
}
\description{
The process involves estimating parameters for a multivariate linear mixed-effects
model, which simultaneously analyzes multiple dependent variables that may be
correlated. This approach incorporates both fixed effects, which are consistent
across the population, and random effects, accounting for variations within
groups or subjects. By fitting this model, one can assess the influence of
predictor variables on several longitudinal outcomes while considering the inherent
variability in the data due to random effects.
}
\examples{

LongSubFixed = list(
  "long1" = serBilir ~ year + age + sex +  (years) + (years) * year,  
  "long2" = prothrombin ~ year + age + sex + (years) + (years) * year,  
  "long3" = albumin ~ year + age + age * year + sex + (years) + (years) * year,  
  "long4" = alkaline ~ year + age + sex + (years) + (years) * year, 
  "long5" = SGOT ~ year + age + sex + (years) + (years) * year, 
  "long6" = platelets ~ year + age + sex + (years)  + (years) * year)

LongSubRandom =list(
  "long1" =  ~ year| id,   
  "long2" =  ~ year| id,    
  "long3" =  ~ year| id,    
  "long4" =  ~ year| id,    
  "long5" =  ~ year| id,    
  "long6" =  ~ year| id)

survivalVariableAll = list(
  "Tyears1",  "Tyears2", "Tyears3", "Tyears4"
)

survivalTransFunction = list(
  fun1 = function(x){abs(x - 1)}, 
  fun2 = function(x){abs(x - 3)}, 
  fun3 = function(x){abs(x - 5)}, 
  fun4 = function(x){abs(x - 7)}
)

# Complete case analysis
data.fit.all = list()
for(i in 1:length(LongSubFixed)){
  data.fit.all[[i]] = pbc3[pbc3$status3 == 1, ]
}

# fitting longitudinal submodel
long_fit_all = longitudinalSub(data.fit.all, LongSubFixed, LongSubRandom)

}
